header: start of the header generator

+++ lib/GlobalVars/GlobalVars.h
@@ -1,6 +1,7 @@
 //De globale variable
 #pragma once
 #include <arduino.h>
+#include <string.h>
 #define no_light 1
 #define blind 2
 #define reflector 3
@@ -13,7 +14,7 @@
 struct headerStruct {
     int datagram_marker;
     int protocol;               
-    int version;
+    double version;
     int length;
     int fragment_offset;
